sync. Keep enrichment ids under controll
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@1806 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ physics/homogenization/elastic_bounds.cpp
@@ -20,7 +20,7 @@ namespace Mu
 
 
 
-ElasticBoundsScheme::ElasticBoundsScheme(size_t i) : Scheme(i)
+ElasticBoundsScheme::ElasticBoundsScheme(int i) : Scheme(i)
 {
 	input.push_back(TAG_VOLUME_FRACTION) ;
 	input.push_back(TAG_BULK_MODULUS) ;
